+++ controllers/courseCtrl.js
@@ -47,12 +47,24 @@ const getCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 const addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
+
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
   if (!bootcamp) {
     return next(
       new ErrorResponse(
         `No bootcamp found with the id of ${req.params.bootcampId}`
+      )
+    );
+  }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add course to this bootcamp ${bootcamp._id}`,
+        404
       )
     );
   }
@@ -69,14 +81,26 @@ const addCourse = asyncHandler(async (req, res, next) => {
 // @route PUT /api/v1/courses/:id
 // @access Private
 const updateCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let course = await Course.findById(req.params.id);
 
   if (!course) {
     return next(new ErrorResponse("No course found with this ID", 404));
   }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add course to update course ${course._id}`,
+        404
+      )
+    );
+  }
+
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
 
   res.status(200).json({
     success: true,
@@ -88,11 +112,23 @@ const updateCourse = asyncHandler(async (req, res, next) => {
 // @route DELETE /api/v1/courses/:id
 // @access Private
 const deleteCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndDelete(req.params.id);
+  let course = await Course.findById(req.params.id);
 
   if (!course) {
     return next(new ErrorResponse("No course found with this ID", 404));
   }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        404
+      )
+    );
+  }
+
+  course = await Course.findByIdAndDelete(req.params.id);
 
   res.status(200).json({
     success: true,
